requirements.txt
selenium==4.24.0
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7
openai==1.5.0
beautifulsoup4==4.12.2